Cimply ... I've given some thought as to the direction
... the main legal avenue I'm persuing is to appeal (as far as possible) the [concurrent use](https://en.wikipedia.org/wiki/Concurrent_use_registration) provision) but **not** by geography

> Concurrent registrations may also be issued by the Director when a court of competent jurisdiction has finally determined that more than one person is entitled to use the same or similar marks in commerce. 
So you can have say |Le⚔DAO for paidup members or |Le✂DAO for affiliated members or |Le⚒️DAO for infostructure layers or even |Le✗DAO for those past lvl3 aiming for an exit. Oh and LeDAO for those journey(wo)men seconded to other DAOs like TEC for cross-training

The legal requirements are
a) court = of which by moral suasion I assert will be the LEETH holders;
b) competent jurisdiction means the squabble of legal engineers need to come to some skill-tree consensus; and 
c) automation of the secondary registry under LexDAO control so no propriated acuity (cf stolen valor) ... ie people passing themselves off as legal engineers when not certified.

The hard part is to negotiate on the co-regulation aspect to give maximal effect to smart contract:
> In issuing concurrent registrations, the Director shall prescribe conditions and limitations as to the mode or place of use of the mark or the goods on or in connection with which such mark is registered to the respective persons.