Update README.md
Cleaning the main constitution Readme.md to contain active and approved constitution and corporate status.  Any changes to this language and repository should be discussed at large with the operating committee in a separate work stream/repository before being merged here.  I've added a comment section at the bottom which will house current thoughts that are not fully incorporated until we determine a working policy for how to manage this repository.